requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langgraph
python-dotenv
pydantic
plotly
PyPDF2
pandas
crewai[litellm]
crewai-tools
litellm
tavily-python
